Stage 091_datasetsRucio
Added:
- comments in datasets_processing.py script
- README file
- input / output examples files
