+++ Utils/Dataflow/091_datasetsRucio/README
@@ -1,0 +1,60 @@
+Stage 091_datasetsRucio
+-----------------------
+
+This stage is implemented to obtain output datasets metadata from Rucio.
+
+The INPUT data for this stage comes from stage 009_oracleConnector.
+
+example:
+{
+	"campaign": "MC15",
+	"processed_events": 395000,
+	"energy_gev": 13000,
+	"site": null,
+	"trans_uses": "Atlas-20.7.5",
+	"geometry_version": "ATLAS-R2-2015-03-01-00",
+	"taskname": "mc15_13TeV.343507.aMcAtNloPythia8EvtGen_A14NNPDF23LO_bbA360_yb2_Zhllbb.recon.e4741_a766_a821",
+	"subcampaign": "MC15c",
+	"job_config": null,
+	"cloud": "WORL",
+	"core_count": "8",
+	"conditions_tags": "OFLCOND-MC15c-SDR-09",
+	"trans_path": "Reco_tf.py",
+	"trans_home": "AtlasProd1-20.7.5.1.1",
+	"evgen_job_opts": null,
+	"ticket_id": "ATLPSTASKS-72836",
+	"taskid": 8343552,
+	"input": ["mc15_13TeV:mc15_13TeV.361034.Pythia8EvtGen_A2MSTW2008LO_minbias_inelastic_low.merge.HITS.e3581_s2578_s2195/", "mc15_13TeV:mc15_13TeV.361035.Pythia8EvtGen_A2MSTW2008LO_minbias_inelastic_high.merge.HITS.e3581_s2578_s2195/", "mc15_13TeV:mc15_13TeV.343507.aMcAtNloPythia8EvtGen_A14NNPDF23LO_bbA360_yb2_Zhllbb.simul.HITS.e4741_a766_tid07402029_00"],
+	"user_name": "gingrich",
+	"architecture": "x86_64-slc6-gcc49-op",
+	"status": "done",
+	"description": "MC15c Higgs BSM other 1",
+	"trigger_config": "RDOtoRDOTrigger=MCRECO:DBF:TRIGGERDBMC:2046,20,5",
+	"start_time": "05-05-2016 12:26:24",
+	"phys_category": [],
+	"vo": "atlas",
+	"task_timestamp": "09-05-2016 13:32:26",
+	"run_number": "343507",
+	"pr_id": 7345,
+	"requested_events": 395000,
+	"project": "mc15_13TeV",
+	"step_name": "Atlfast",
+	"phys_group": "MCGN",
+	"output": ["mc15_13TeV.343507.aMcAtNloPythia8EvtGen_A14NNPDF23LO_bbA360_yb2_Zhllbb.recon.AOD.e4741_a766_a821_tid08343552_00"],
+	"hashtag_list": "aMcAtNloPythia8EvtGen",
+	"end_time": "09-05-2016 11:18:32"
+}
+
+Stage 091_datasetsRucio constructs the following NDJSON string, with 'bytes' from Rucio:
+
+example:
+{
+	"output": [{
+		"deleted": true,
+		"datasetname": "mc15_13TeV.343507.aMcAtNloPythia8EvtGen_A14NNPDF23LO_bbA360_yb2_Zhllbb.recon.AOD.e4741_a766_a821_tid08343552_00",
+		"bytes": -1
+	}],
+	"taskid": 8343552
+}
+
+
